// src/enums/CatalogFilter.ts
export enum CatalogFilter {
    Sort = 'sort',
    Category = 'category',
    Color = 'color',
    Price = 'price',
    Room = 'room',
    Collection = 'collection',
    Sizes = 'sizes',
    Sale = 'sale',
}

export enum SizesFilter {
    Width = 'width',
    Height = 'height',
    Length = 'length',
    Depth = 'depth',
}